requirements.txt
livekit-agents>=1.2.1
livekit-plugins-openai>=1.2.1
livekit-plugins-noise-cancellation>=0.2.5
python-dotenv>=1.1.0
livekit>=1.0.12
livekit-api>=1.0.3
livekit-protocol>=1.0.4 